chore: extract common zerion types
