+++ packages/types/src/index.ts
@@ -1,3 +1,4 @@
 export * from './base'
 export * from './market'
 export * from './utility'
+export * from './zerion'
