+++ packages/types/src/zerion.ts
@@ -1,0 +1,43 @@
+import type { ChainId } from '@shapeshiftoss/caip'
+import {
+  arbitrumChainId,
+  avalancheChainId,
+  baseChainId,
+  bscChainId,
+  ethChainId,
+  gnosisChainId,
+  optimismChainId,
+  polygonChainId,
+} from '@shapeshiftoss/caip'
+
+export const ZERION_CHAINS = [
+  // shapeshift supported
+  'avalanche',
+  'binance-smart-chain',
+  'ethereum',
+  'optimism',
+  'polygon',
+  'arbitrum',
+  'xdai',
+  'base',
+  // not yet
+  // 'aurora',
+  // 'fantom',
+  // 'solana',
+] as const
+
+export type ZerionChainId = (typeof ZERION_CHAINS)[number]
+
+export const ZERION_CHAINS_MAP: Record<ZerionChainId, ChainId> = {
+  avalanche: avalancheChainId,
+  'binance-smart-chain': bscChainId,
+  ethereum: ethChainId,
+  optimism: optimismChainId,
+  polygon: polygonChainId,
+  arbitrum: arbitrumChainId,
+  xdai: gnosisChainId,
+  base: baseChainId,
+}
+
+export const zerionChainIdToChainId = (chainId: ZerionChainId): ChainId | undefined =>
+  ZERION_CHAINS_MAP[chainId]
