+++ scripts/generateAssetData/generateRelatedAssetIndex/mapping.ts
@@ -1,50 +1,8 @@
-import type { AssetId, ChainId } from '@shapeshiftoss/caip'
-import {
-  arbitrumChainId,
-  ASSET_NAMESPACE,
-  avalancheChainId,
-  baseChainId,
-  bscChainId,
-  ethChainId,
-  gnosisChainId,
-  optimismChainId,
-  polygonChainId,
-  toAssetId,
-} from '@shapeshiftoss/caip'
+import type { AssetId } from '@shapeshiftoss/caip'
+import { ASSET_NAMESPACE, bscChainId, toAssetId } from '@shapeshiftoss/caip'
+import { type ZerionChainId, zerionChainIdToChainId } from '@shapeshiftoss/types'
 
 import type { ZerionImplementation } from './validators/fungible'
-
-export const ZERION_CHAINS = [
-  // shapeshift supported
-  'avalanche',
-  'binance-smart-chain',
-  'ethereum',
-  'optimism',
-  'polygon',
-  'arbitrum',
-  'xdai',
-  'base',
-  // not yet
-  // 'aurora',
-  // 'fantom',
-  // 'solana',
-] as const
-
-export type ZerionChainId = (typeof ZERION_CHAINS)[number]
-
-export const ZERION_CHAINS_MAP: Record<ZerionChainId, ChainId> = {
-  avalanche: avalancheChainId,
-  'binance-smart-chain': bscChainId,
-  ethereum: ethChainId,
-  optimism: optimismChainId,
-  polygon: polygonChainId,
-  arbitrum: arbitrumChainId,
-  xdai: gnosisChainId,
-  base: baseChainId,
-}
-
-export const zerionChainIdToChainId = (chainId: ZerionChainId): ChainId | undefined =>
-  ZERION_CHAINS_MAP[chainId]
 
 export const zerionImplementationToMaybeAssetId = (
   implementation: ZerionImplementation,
